Relocate code and fix types
